Move new types to `types/api`
