+++ frontend/src/metabase-types/api/user.ts
@@ -1,7 +1,15 @@
 export interface User {
   id: number;
+  common_name: string;
   first_name: string;
+  last_name: string;
+  email: string;
+  google_auth: boolean;
+  is_active: boolean;
+  is_qbnewb: boolean;
   is_superuser: boolean;
+  date_joined: string;
+  last_login: string;
   has_invited_second_user: boolean;
   personal_collection_id: number;
 }
